requirements.txt
Django==3.2.2
djangorestframework==3.13.1
djangorestframework_simplejwt==4.6.0
PyJWT==2.3.0
